+++ src/__test__/calc.test.js
@@ -1,0 +1,7 @@
+import { Calc } from '..';
+
+describe('calc.js', () => {
+	test('验证api', () => {
+		expect(Calc(1).add(1).val()).toBe(2);
+	});
+});