+++ src/__test__/device.test.js
@@ -1,0 +1,7 @@
+import { Device } from '..';
+
+describe('device.js', () => {
+	test('验证api', () => {
+		expect(typeof Device).toBe('object');
+	});
+});