+++ src/__test__/utils.test.js
@@ -1,0 +1,7 @@
+import { Utils } from '..';
+
+describe('url.js', () => {
+	test('验证api', () => {
+		expect(typeof Utils).toBe('object');
+	});
+});