+++ src/funcs/assit.js
@@ -1,7 +1,7 @@
 /**
  * 小于10的数字前面加0
  */
-export const prefixZero = (num) => {
+export const preZero = (num) => {
 	if (num < 10 && num > 0) {
 		return "0" + num;
 	} else if (num <= 0) {