+++ src/funcs/index.js
@@ -1,0 +1,24 @@
+import {
+	preZero,
+	def,
+	isObj,
+	hasOwn,
+	cloneDeep,
+	cloneDeepEasier
+} from './assit';
+import { createMixins } from './mixin';
+import { formatMoney } from './format';
+import { canvas2file, getWordsLength } from './other';
+
+export const Utils = {
+	preZero,
+	def,
+	isObj,
+	hasOwn,
+	cloneDeep,
+	cloneDeepEasier,
+	createMixins,
+	formatMoney,
+	canvas2file,
+	getWordsLength
+};