+++ src/funcs/mixin.js
@@ -1,0 +1,6 @@
+// 用于对象 
+// @createMixins({})
+// class {}
+export const createMixins = (...mixins) => target => {
+	Object.assign(target.prototype, ...mixins);
+};