+++ src/index.js
@@ -1,0 +1,8 @@
+export { Stroage, Cookie } from './cache';
+export { URL } from './url';
+export { RegEx } from './regex';
+export { Device } from './device';
+export { Calc } from './calc';
+export { Utils } from './funcs/index';
+
+